pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "badoo-bot"
version = "0.1.0"
description = "Badoo tools for enhace badoo experience with automationa"
authors = [{ name = "joaog", email = "joaogmendess19@gmail.com" }]
requires-python = ">=3.10"
dependencies = ["requests"]

[project.optional-dependencies]

dev = ["pytest", "pytest-playwright", "black", "isort", "flake8", "mypy"]


[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.flake8]
max-line-length = 88
exclude = ".venv,build,dist,__pycache__,.git"
extend-ignore = "E203,W503"

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["build", "dist", "tests/helpers"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
